refactor: wkc layout — phase 17 (collapse last legacy src/ folders)
after phase 16, two single-file legacy directories survived under
src/. neither hosted a wkc component; both held helpers used by
exactly one caller. relocates them so the only top-level src/
directories are the wkc-shaped ones (adapters/, logic/, controllers/,
entrypoints/, migrations/).

- src/service/time/TimeSorting.ts → src/logic/deployment-service/
  time-sorting.ts. the only consumer is deployment-service. moving
  it into that component's folder keeps the file collocated with its
  caller; renamed to kebab-case to match the rest of the folder
  (component.ts, types.ts, index.ts).
- src/helpers/RetryHelper.ts → test/helpers/RetryHelper.ts. the only
  consumer is E2ETestUtils.ts. it is a generic retry-on-throw helper
  used solely by integration tests, so it belongs under test/helpers/
  alongside PromiseAssertions.ts (filename kept PascalCase to match
  that file).

with both files relocated, src/service/ and src/helpers/ are empty
and are removed. content/src/ now contains only the wkc layout plus
the bootstrap files (Environment.ts, metrics.ts, instrument.ts,
deployment-types.ts, types.ts, components.ts, service.ts).

yarn build is clean. 204 tests pass across the affected paths
(deployer, delete-unreferenced-files, snapshots).
